4auth_service/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
databases[postgresql]==0.9.0  # 升级到支持 SQLAlchemy 2.0 的版本
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psycopg2-binary